+++ src/prodigy_prot/__init__.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+NACCESS_CONFIG = Path(Path(__file__).parents[0], "data/naccess.config")
